requirements.txt
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.23
langchain-chroma==0.2.4
langchain-huggingface==0.1.2
chromadb==1.0.9
openai==1.70.0
python-dotenv==1.0.1
streamlit==1.44.1
pypdf==4.3.1
tiktoken==0.9.0
sentence-transformers==3.4.1